robot/requirements.txt
websockets
httpx
python-dotenv
Pillow
numpy
pygame
pyttsx3
reachy-mini
